rename the preload folder to loadable to reflect the exported path
