+++ packages/sku/src/services/vite/loadable/helpers/scriptUtils.ts
